ready to debbug angle

+++ includes/cub3d.h
@@ -6,7 +6,7 @@
 /*   By: pkorsako <pkorsako@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/20 17:15:52 by pkorsako          #+#    #+#             */
-/*   Updated: 2024/03/13 18:01:24 by ajeannin         ###   ########.fr       */
+/*   Updated: 2024/03/13 19:43:36 by ajeannin         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -77,8 +77,8 @@ typedef struct s_grid
 
 typedef struct s_player
 {
-	int		pos_x;
-	int		pos_y;
+	double		pos_x;
+	double		pos_y;
 	double	orientation;
 	int		x_max;
 	int		y_max;
@@ -135,13 +135,13 @@ unsigned int	proj_plan_col(t_game *game, double angle);
 double			dtor(double degrees);
 double			rtod(double radian);
 double			double_modulo(double value, double modulo);
-int				ft_dist(int x1, int x2, int y1, int y2);
+int				ft_dist(double x1, double x2, double y1, double y2);
 int				get_value(unsigned int value, const char *flag);
 void			view_stocked_image(unsigned int *stock);
 void			view_stocked_col(unsigned int stock);
 void			max_dist(t_game *game);
-int				get_ipx(t_player *player, int i_py, double angle);
-int				get_ipy(t_player *player, int i_px, double angle);
+int				get_ipx(t_player *player, double i_py, double angle);
+int				get_ipy(t_player *player, double i_px, double angle);
 char			try_get_texture(char **map, int yt, int xt);
 
 //////////// Maths_tests ///////////////////
